Split CAT_database.py into database.py and database_functions.py
The Database class is contained within database.py
All remaining functions are in database_functions.py
