+++ CAT/analysis/ligand_bde.py
@@ -19,7 +19,7 @@ from .jobs import (job_single_point, job_geometry_opt, job_freq)
 from ..utils import (get_time, type_to_string)
 from ..mol_utils import (to_atnum, merge_mol)
 from ..attachment.ligand_attach import rot_mol_angle
-from ..data_handling.CAT_database import Database
+from ..data_handling.database import Database
 
 
 def init_bde(qd_df, arg):
