+++ CAT/attachment/ligand_attach.py
@@ -9,9 +9,10 @@ from scipy.spatial.distance import cdist
 from scm.plams.mol.molecule import Molecule
 from scm.plams.core.settings import Settings
 
-from ..utils import (get_time, type_to_string)
+from ..utils import get_time
 from ..mol_utils import (merge_mol, get_atom_index)
-from ..data_handling.CAT_database import (Database, mol_to_file)
+from ..data_handling.database import Database
+from ..data_handling.database_functions import mol_to_file
 
 
 def init_qd_construction(ligand_df, core_df, arg):
