+++ CAT/data_handling/__init__.py
@@ -1,12 +1,14 @@
 """ Modules related to the importing, exporting and general handling of data. """
 
-from .CAT_database import (Database, mol_to_file)
+from .database import Database
+from .database_functions import mol_to_file
 from .mol_import import (read_mol, set_mol_prop)
 from .input_sanitizer import (sanitize_optional, sanitize_input_mol, sanitize_path)
 
 
 __all__ = [
-    'Database', 'mol_to_file',
+    'Database',
+    'mol_to_file',
     'read_mol', 'set_mol_prop',
     'sanitize_optional', 'sanitize_input_mol', 'sanitize_path'
 ]
